ARC: disable ioremap_prot for HS5x
Function ioremap_prot uses unsigned long type for flags. However
ARCv3 HS5x CPU family, being 32-bit platform, requires 64-bit page
protocol flags. Disable IOREMPA_PROT for now.

+++ arch/arc/include/asm/page.h
@@ -110,7 +110,7 @@ typedef struct {
 
 typedef struct {
 #ifndef CONFIG_ARC_MMU_V6
-	unsigned long pgprot;
+	unsigned long long pgprot;
 #else
 	unsigned long long pgprot;
 #endif
